Split Bcrypt and OTP into separate modules
